Part: Fix regression in TopoShape::makeElementCopy
See https://forum.freecad.org/viewtopic.php?t=94158 for an example where applying the transformation from source
to the copy leads to a wrong shape. So, the change must be reverted.

The change originally was added by PR 17436 to fix an issue with ruled surfaces. By reverting the change the issue
must be rechecked and a more reliable fix is needed.