Part: Revert questionable change in Part::Feature::onChanged
The change was made with PR 13964 but a reason for its necessity was never delivered. However, it's known to cause a serious
performance issue as described in the forum thread https://forum.freecad.org/viewtopic.php?t=92201.
An analysis of the exact problem is given here https://forum.freecad.org/viewtopic.php?p=793096#p793096

By reverting the change the animation script works again as smooth as in v.0.21.
Also, reverting the change shows no regression in any of the unit tests.