PD: Fix loft between curved faces
This fixes issue 19183. It's also reported in the forum https://forum.freecad.org/viewtopic.php?p=806495 and
https://forum.freecad.org/viewtopic.php?t=88234.

The actual reason for the failure is that the new code uses the wrong face maker method (FaceMakerBullseye)
that doesn't support curved wires while in the old code FaceMakerCheese was used that supports it.

So, changing the face maker method fixes the problem.