PD: Fix regression about single-solid check
For several PD features the single-solid check fails. The regression is caused by PR 13960.
The reason for the failure is that the first solid of the output shape is retrieved and then
checked for a single solid. This test will always pass, of course.

The single-solid is fixed for these features:
* Pad
* Pocket (never worked there)
* Fillet
* Chamfer
* Groove (never worked there)
* Revolution (never worked there)
* Loft

This fixes issue 19002