Sketch: Fix crash after creating B-Spline
If the B-Spline handler is not in continuous mode it will be deleted after the creation of the B-spline.
This happens inside DrawSketchHandlerBSpline::quit() by the call of 'setState(SelectMode::End)'. After returning
from this function 'finish()' mustn't be called any more as otherwise a segmentation fault will occur.

The solution is to check beforehand if the handler is in continuous mode and if not only call 'setState(SelectMode::End)'.

This fixes issue 18660