Gui: Fix DlgPreferencesImp::minimumDialogWidth() to return a reasonable width
The current implementation always returns a width that is too small so that a scrollbar is shown. Especially on the General page this is unfortunate
as some important controls are truncated or completely hidden.

Another problem of this implementation is that when loading a not yet loaded workbench in the 'Available workbenches' page then the computed size
will be by far too high leading to a dialog that all the sudden covers most of the screen.

Solution:
The correct width is the sum of: width of the tree view + width of the biggest page + spacing of the layout
This fixes both of the above problems.