Part: Adjust Mirroring::execute()
By fixing TopoShape::makeElementMirror() the method Mirroring::execute() must be adjusted too because
it doesn't handle the placement of the input shape any more.

There are two ways to do it:
1. Set Placement property of the Mirror feature to the placement of the input feature.
2. Apply the placement to transformation matrix that will be applied to the input shape.

Because in all older FreeCAD versions the option 2. was implemented it will be done here, too.

This change is a reimplementation of the commit b3156069a from Dec 13, 2011 to fix the bug
'0000531: Python API: Mirroring rotated parts produces undesired results' on Mantis