Gui: Fix shading after leaving a sketch from edit mode
If the body of a sketch is made invisible then leaving the sketch from edit mode fails to restore the original shading settings.
This is because View3DInventorViewer::searchNode() fails to find the passed Inventor node. The reason is that if an SoSwitch has
set its whichChild field to SO_SWITCH_NONE its children won't be considered for the search.

To fix the problem call setSearchingAll(true) of the SoSearchAction object.

This fixes issue 19338