Measure: Fix several crashes in Measurement
Measurement::getShape() used to call ShapeFinder::getLocatedShape() that failed to return a valid shape in many cases.
And the calling instance didn't check for a valid shape either which caused a segmentation fault.

Solution:
Replace ShapeFinder::getLocatedShape() with Part::Feature::getShape() that works more reliably and where needed check
for a valid shape before accessing it.

This fixes issue 19452 and improves the fix for 19349