Part: Fix Extrusion::calculateShapeNormal
This method makes the wrong assumption that a planar face of a Part2DObject always has (0, 0, 1) as normal.
Hence, it fails to compute the correct extrusion direction.

Solution:
Also use BRepLib_FindSurface as for the general case and use the old code as fallback in case the Part2DObject
hasn't a plane, e.g. if it's a single line.

Hint: The same wrong assumption is made in ProfileBased::getProfileNormal() in the PartDesign module.

This fixes the issue as desribed in the forum thread: https://forum.freecad.org/viewtopic.php?t=94646