TD: Fix regression in DrawViewSection::makeSectionCut
There is no reason why FCBrepAlgoAPI_Cut should be preferred over BRepAlgoAPI_Cut in this function.
In the error analysis it's said that also BRepAlgoAPI_Cut may throw an exception. So, therefore just in case a lambda
function is implemented with a try/catch block to avoid that it leaves the for loop too early.
Inside the lambda function BRepAlgoAPI_Cut is used again because it may still return a result where FCBrepAlgoAPI_Cut
unnecessarily throws an exception.

This fixes issue 17875