PD: Improve groove feature
The only difference between groove and revolution is that for the former the
revolved face is removed from the base shape instead of added.

Thus, the code of the Revolution and Groove classes should be almost identical.
This allows it in a further step to refactor the code and make a common base class.

This fixes issue 18842.