TD: Fix crash on selecting 3D edge for dimension
Unlike in the bug description of issue 19654 the user doesn't have to preselect an edge in the 3D view
but first start the dimension command and then select an edge.

This commit adds some security checks to TechDrawHandler::quit(), TechDrawHandler::getPage() and
TDHandlerDimension::onSelectionChanged() to be on the safe side that no null pointers are dereferenced.

But the ultimative fix for this whole problem is to change activateHandler() and immediately delete
the passed TechDrawHandler if it fails to find the appropriate QGVPage. This is needed as otherwise the
handler behaves like a ghost object that affects the selection mechanism and disallows to select anything
in the 3D view or the tree view.

Fixes issue 19654