Part: Fix slice feature
If the Base link points to a group object an exception is raised that it doesn't have a Shape attribute.

Solution:
Explicitly check for the Shape and Group attributes. In the former case use the shape for further processing.
In the latter case access the objects of the Group and create a Part compound from their shapes. Pass the compound
for further processing.

This fixes issue 19873