Sheet: Disable zooming
With PR 16130 the zooming capability was added to spreadsheet view.

The SheetTableView is added to a QGraphicsScene of a QGraphicsView that provides the method scale().
However, this causes some problems with header of the table view and makes resizing columns or rows
very unintuitive.

A correctly working table view is probably more important than the possibility to zoom in or out and that's
why the latter will be disabled.

This fixes issue 19863