App: Improve Document::writeObjects
Do not directly pass getNameInDocument() to the ostream as this will set the badbit in case it returns nullptr.
As a result no further output is written to the stream and leads to loss of data.

Instead the returned C string is stored in a local variable and in case it's nullptr an empty string is passed
to the ostream.

At the end of the function writeObjects() it's checked whether the failbit or badbit is set. If yes an error
message is printed and the failbit or badbit are cleared.

This is supposed to fix the issue 18044.

In case this fix is not sufficient it can be also check for every object inside writeObjectData()