TD: Fix loading project files
With PR 19941 there is a regression where parts of the document won't be restored. This is caused by
overriding the virtual method handleChangedPropertyType() but not calling the method of the base class.

While searching for a fix another problem showed up where (due to the incomplete loading) exceptions are
thrown but not handled within TD. Therefore the method recompute() is overridden in DrawView to also handle
OCC exceptions.