Measure: Extend area measurements
Currently the measurement of areas is limited to cylinders and planes, i.e. it's not possible to measure the area of spheres, toruses or B-spline surfaces.
To support further surface types the new enum value 'SURFACE' is added and used in the PartMeasureTypeCb callback. And the class MeasureArea explicitly allows
any surface type now.

This fixes issue 20802